requirements.txt

torch
torchvision
numpy
Pillow
xgboost
flask
scikit-learn